requirements.txt
annotated-types==0.5.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
attrs==22.2.0
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
click==8.1.6
colorama==0.4.6
commonmark==0.9.1
constantly==15.1.0
contextlib2==21.6.0
contourpy==1.1.0
cryptography==41.0.3
cssselect2==0.7.0
cycler==0.11.0
Cython==3.0.2
deprecation==2.1.0
distlib==0.3.7
distro==1.8.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.20.1
et-xmlfile==1.1.0
evdev==1.6.1
exceptiongroup==1.1.3
fastjsonschema==2.18.0
filelock==3.12.2
Flask==2.2.5
fonttools==4.42.1
future==0.18.3
gmpy2==2.1.5
gotrue==1.0.4
greenlet==2.0.2
h11==0.14.0
hid-parser==0.0.3
html5lib==1.1
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==5.0.0
incremental==22.10.0
inflect==7.0.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jaraco.context==4.3.0
jaraco.functools==3.8.1
jaraco.text==3.11.1
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==24.2.0
kiwisolver==1.4.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
material-color-utilities-python==0.1.5
matplotlib==3.7.2
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.1
npyscreen==4.10.5
nspektr==0.4.0
ordered-set==4.1.0
packaging==23.1
postgrest==0.10.8
pyOpenSSL==23.2.0
pyproject_hooks==1.0.0
PyQt5==5.15.9
PyQt5-sip==12.12.2
pyrsistent==0.19.3
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-xlib==0.33
realtime==1.0.0
referencing==0.30.2
regex==2023.6.3
reportlab==3.6.12
requests==2.28.2
resolvelib==1.0.1
supabase==1.0.4
supafunc==0.2.3
svglib==1.5.1
sympy==1.12
tenacity==8.2.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
Twisted==22.10.0
typing_extensions==4.7.1
uc-micro-py==1.0.2
urllib3==1.26.15
userpath==1.9.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.2
Yapsy==1.12.2
zipp==3.16.2
zope.interface==6.0
